strands-agents-examples/90-deployment/10-serverless/10-weather-agent-api/agent/requirements.txt
strands-agents==1.13.0
strands-agents-tools==0.2.12
python-json-logger==4.0.0
